add support for drawer item
